Combine heex and elixir modes into one file
heex-ts-mode requires elixir-ts-mode and vice versa. unless there is a
trick to make it work without duplicating many of the major mode
functions it seems to make sense to combine it for now.

Another option would be to have a "light" version of elixir in
heex-ts-mode as we really just need syntax highlighting.